started with the boiler plate
now i have a node backend and a react front end.  I have demonstrated sending informaiton back and forth between the client and the server.

there is still a good amount to do.
 there needs to be a qeustion input form that is only visible to the broadcaster (or mods?)

and then users need to be able to vote.

once this is the case, there needs to be a pre and post vote screen.

the server needs to send out a broadcast to show the updated quesion stats.